Created ButtonBlock component for optimization
