+++ src/lib/index.ts
@@ -2,5 +2,6 @@
 import Counter from "./components/Counter.svelte";
 import CounterList from "./components/CounterList.svelte";
 import Dashboard from "./components/Dashboard.svelte";
+import ButtonBlock from "./components/ButtonBlock.svelte";
 
-export { Counter, CounterList, Dashboard };
+export { Counter, CounterList, Dashboard, ButtonBlock };
